refactor: 북마크 저장, 해제 리팩토링, 예외처리 #19
